requirements.txt
yfinance
pandas
scikit-learn
nltk 
newsapi-python
pandas-ta
xgboost